refactor(chart-utils): Refactor
- Moved chart utils to separate file from the hook
